Fix build issues with InfluxDBv1
It seems somebody moved the v1.11.4 tag in the influxdata/influxdb
repository after a recent bugfix. That's a really bad thing to do and it
broke the go module dependency handling.
As it turns out, the InfluxDBv1 client was split off into a standalone
repository back in 2019, so switch to using that instead.