Fix handling of pinned workloads for Prometheus
Pinned workflows show up as a non-null workload type, but they are not
an overflow bucket unlike the other workload types.
The Prometheus code checked the workload type because it has to
incorporate it into the mertic name (because Prometheus doesn't support
multivalued metrics) and dropped the metric because "Pinned" was an
unknown workload type.
The fix treats the metric as a normal metric (not an overflow), but the
change adds an extra label "pinned". This means that your metric will
change since it is defined by the name and the complete set of labels.